chore: migrate run utils to Typescript
